@@ -23,22 +23,22 @@
     "core-js": "^3.2.1",
     "geojson": "^0.5.0",
     "lodash": "^4.17.19",
-    "mapbox-gl": "^1.3.1",
+    "mapbox-gl": "^1.13.0",
     "normalize.css": "^8.0.1",
     "prop-types": "^15.6.2",
     "qs": "^6.9.4",
-    "react": "^16.10.1",
-    "react-autosuggest": "^9.4.3",
+    "react": "^16.14.0",
+    "react-autosuggest": "^10.0.4",
     "react-dom": "^16.10.1",
-    "react-modal": "^3.11.2",
+    "react-modal": "^3.12.1",
     "react-number-format": "^4.4.1",
     "react-router-dom": "^5.2.0",
-    "react-transition-group": "^4.3.0",
+    "react-transition-group": "^4.4.1",
     "react-virtualized-auto-sizer": "^1.0.2",
     "react-window": "^1.8.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.0",
+    "@babel/core": "^7.12.10",
     "@babel/plugin-proposal-class-properties": "^7.5.5",
     "@babel/preset-env": "^7.6.0",
     "@babel/preset-react": "^7.0.0",
@@ -55,12 +55,12 @@
     "@types/geojson": "^7946.0.7",
     "@types/gtag.js": "0.0.3",
     "@types/html-webpack-plugin": "^3.2.3",
-    "@types/less": "^3.0.1",
-    "@types/lodash": "^4.14.149",
-    "@types/mapbox-gl": "^0.54.3",
-    "@types/mini-css-extract-plugin": "^0.8.0",
+    "@types/less": "^3.0.2",
+    "@types/lodash": "^4.14.165",
+    "@types/mapbox-gl": "^1.12.8",
+    "@types/mini-css-extract-plugin": "^1.2.2",
     "@types/modernizr": "^3.5.3",
-    "@types/prettier": "^1.19.1",
+    "@types/prettier": "^2.1.5",
     "@types/prop-types": "^15.7.3",
     "@types/qs": "^6.9.4",
     "@types/react": "^16.9.2",
@@ -72,45 +72,44 @@
     "@types/react-virtualized-auto-sizer": "^1.0.0",
     "@types/react-window": "^1.8.2",
     "@types/uglifyjs-webpack-plugin": "^1.1.0",
-    "@types/webpack": "^4.41.21",
+    "@types/webpack": "^4.41.25",
     "@types/webpack-bundle-analyzer": "^3.8.0",
     "@types/webpack-dev-server": "^3.11.0",
     "@types/webpack-merge": "^4.1.5",
     "@typescript-eslint/eslint-plugin": "^2.3.0",
-    "@typescript-eslint/parser": "^2.3.0",
-    "autoprefixer": "^9.6.1",
+    "@typescript-eslint/parser": "^2.34.0",
+    "autoprefixer": "^9.8.6",
     "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "copy-webpack-plugin": "^6.0.3",
-    "css-loader": "^3.2.0",
+    "babel-loader": "^8.2.2",
+    "copy-webpack-plugin": "^6.4.0",
+    "css-loader": "^5.0.1",
     "eslint": "^6.4.0",
     "eslint-config-prettier": "^6.3.0",
     "eslint-config-squarespace": "^1.0.1",
-    "eslint-import-resolver-webpack": "^0.11.1",
+    "eslint-import-resolver-webpack": "^0.13.0",
     "eslint-loader": "^4.0.2",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.10.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-react": "^7.21.5",
     "expose-loader": "^0.7.5",
-    "file-loader": "^6.0.0",
+    "file-loader": "^6.2.0",
     "html-webpack-plugin": "^3.2.0",
-    "husky": "^3.0.5",
     "json-loader": "^0.5.7",
     "less": "^3.10.3",
     "less-loader": "^5.0.0",
-    "lint-staged": "^9.3.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "modernizr": "^3.11.3",
+    "lint-staged": "^10.5.3",
+    "mini-css-extract-plugin": "^1.3.3",
+    "modernizr": "^3.11.4",
     "modernizr-loader": "^1.0.1",
-    "postcss-loader": "^3.0.0",
+    "postcss-loader": "^4.1.0",
     "prettier": "^2.2.1",
     "react-hot-loader": "^4.13.0",
     "ts-loader": "^6.1.2",
-    "typescript": "^3.6.3",
+    "typescript": "^4.1.3",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^4.40.2",
-    "webpack-bundle-analyzer": "^3.5.1",
+    "webpack-bundle-analyzer": "^4.3.0",
     "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1",
+    "webpack-dev-server": "^3.11.0",
     "webpack-merge": "^4.1.3"
   },
   "husky": {
@@ -120,8 +119,7 @@
   },
   "lint-staged": {
     "src/**/*.{js,jsx,ts,tsx}": [
-      "eslint --fix --max-warnings 0",
-      "git add"
+      "eslint --fix --max-warnings 0"
     ]
   },
   "files": [
