@@ -1,9 +1,13 @@
-const sharp = require('sharp');
-const AWS = require('aws-sdk');
+import sharp from 'sharp';
+import AWS from 'aws-sdk';
+
+import * as LaserdiscUtils from './src/image-processing/LaserdiscUtils';
 
 const s3 = new AWS.S3();
 
-const FILENAMES = {
+type Template = { prefix: string; suffix: string };
+
+const FILENAMES: Record<string, Template> = {
   jpeg: {
     prefix: 'jpg/',
     suffix: '.jpg',
@@ -14,15 +18,15 @@ const FILENAMES = {
   },
 };
 
-const makeFilename = (template, rootKey) =>
+const makeFilename = (template: Template, rootKey: string): string =>
   template.prefix + rootKey + template.suffix;
 
 const INPUT_PREFIX = 'originals/';
 
-const getRootKey = srcKey =>
+const getRootKey = (srcKey: string): string =>
   srcKey.substring(INPUT_PREFIX.length).replace(/\.\w+$/, '');
 
-module.exports.handler = async event => {
+export const handler = async (event): Promise<unknown> => {
   const srcBucket = event.Records[0].s3.bucket.name;
   const srcKey = event.Records[0].s3.object.key;
 
@@ -40,7 +44,12 @@ module.exports.handler = async event => {
     })
     .promise();
 
-  const inputBuffer = inputObject.Body;
+  let inputBuffer = inputObject.Body as Buffer;
+
+  // Crop laserdisc video frames to eliminate borders and superimposed banner
+  if (await LaserdiscUtils.isLaserdiscVideoFrame(inputBuffer)) {
+    inputBuffer = await LaserdiscUtils.cropVideoFrame(inputBuffer);
+  }
 
   return Promise.all([
     // webp doesn't look that good, sorry webp
@@ -61,7 +70,7 @@ module.exports.handler = async event => {
     sharp(inputBuffer)
       .jpeg()
       .toBuffer()
-      .then(outputBuffer =>
+      .then((outputBuffer) =>
         s3
           .putObject({
             Body: outputBuffer,
@@ -74,10 +83,10 @@ module.exports.handler = async event => {
       ),
 
     sharp(inputBuffer)
-      .resize(420)
+      .resize(420, undefined, { withoutEnlargement: true })
       .jpeg()
       .toBuffer()
-      .then(outputBuffer =>
+      .then((outputBuffer) =>
         s3
           .putObject({
             Body: outputBuffer,
@@ -91,7 +100,7 @@ module.exports.handler = async event => {
   ]);
 };
 
-module.exports.deletionHandler = async event => {
+export const deletionHandler = async (event): Promise<unknown> => {
   const srcBucket = event.Records[0].s3.bucket.name;
   const srcKey = event.Records[0].s3.object.key;
 
@@ -107,8 +116,8 @@ module.exports.deletionHandler = async event => {
       Bucket: srcBucket,
       Delete: {
         Objects: Object.values(FILENAMES)
-          .map(template => makeFilename(template, rootKey))
-          .map(key => ({
+          .map((template) => makeFilename(template, rootKey))
+          .map((key) => ({
             Key: key,
           })),
       },
